refactor: move import function out of `tilia.ui.QtUI`
This keeps the class more cohesive and should facilitate testing.
